Move the diameter arg from base to cellpose cli.
I want to be able to specify another default value for
omnipose.
